fix(modal): click outside modal focus' out, address comments
