+++ __tests__/modal.js
@@ -1,8 +1,7 @@
 import React from 'react';
-import { mount, shallow } from 'enzyme';
+import { mount } from 'enzyme';
 import CSSTransition from 'react-transition-group/CSSTransition';
 import Modal from '../src/modal';
-import ConditionalWrap from '../src/conditional-wrap';
 
 const defaultProps = {
   classes: {
@@ -375,38 +374,5 @@ describe('modal', () => {
       await wait();
       expect(document.documentElement.style.overflow).toBe('');
     });
-  });
-});
-
-describe('ConditionalWrap', () => {
-  it('Should wrap when true', () => {
-    const wrapper = shallow(
-      <ConditionalWrap condition wrap={children => <a href='/'>{children}</a>} >
-        <p>content</p>
-      </ConditionalWrap>
-    );
-
-    expect(wrapper.find('a').length).toBe(1);
-    expect(wrapper.find('p').length).toBe(1);
-  });
-  it('Should not wrap when false', () => {
-    const wrapper = shallow(
-      <ConditionalWrap condition={false} wrap={children => <a href='/'>{children}</a>} >
-        <p>content</p>
-      </ConditionalWrap>
-    );
-
-    expect(wrapper.find('a').length).toBe(0);
-    expect(wrapper.find('p').length).toBe(1);
-  });
-  it('Should not wrap when undefined', () => {
-    const wrapper = shallow(
-      <ConditionalWrap wrap={children => <a href='/'>{children}</a>} >
-        <p>content</p>
-      </ConditionalWrap>
-    );
-
-    expect(wrapper.find('a').length).toBe(0);
-    expect(wrapper.find('p').length).toBe(1);
   });
 });
